This calculator is very basic one and it is designed to experiment on the doherty threshold.
The calculator is designed to display the delay time for the result to display on the screen. It helps us to collect data in the interest of our users. based on users' responses for the different delay times, we can prove that user's interest decreases as the delay time increases.